feat/ajaxified capsule form comportment & added delete on form draft & added prefil to input on edit
