+++ app/views/capsules/edit.html.erb
@@ -1,4 +1,3 @@
 <div data-controller="toggle">
-  <%= render "shared/form_capsule_page_1", local: { capsule: @capsule } %>
-  <%= render "shared/form_capsule_page_2", local: { capsule: @capsule } %>
+  <%= render "shared/form_capsule", local: { capsule: @capsule } %>
 </div>
