+++ app/views/shared/_form_capsule.html.erb
@@ -1,0 +1,65 @@
+<div class="container"&>
+  <%= link_to 'Back', capsules_path,  data: { 'toggle-target': 'capsuleInput' }, method: :patch %>
+  <a data-action="click->toggle#toggleInputs" class="d-none" data-toggle-target="capsuleInput">Back</a>
+  <!-- Toggle message section -->
+  <div class="container mt-3">
+    <% @capsule.messages.each do |message| %>
+      <div class="form-control">
+        <p><%= message.title %></p>
+        <p><%= message.content %></p>
+        <%= link_to 'Edit', edit_capsule_message_path(@capsule, message) %>
+        <%= button_to 'Delete', capsule_message_path(@capsule, message), method: :delete %>
+      </div>
+    <% end %>
+  </div>
+  <!-- Toggle photos section -->
+  <div>
+    <% @capsule.photos.each do |photo| %>
+      <%# <%= render partial: "capsules/thumbnail", locals: { photo: photo } %> %>
+    <% end %>
+  </div>
+  <!-- Capsule form -->
+  <%= simple_form_for @capsule, html: { multipart: true } do |f| %>
+    <%= f.input :name, label: 'Capsule name', wrapper_html: { data: { 'toggle-target': 'capsuleInput' } }, hmtl: { value: @capsule.name } %>
+    <%= link_to 'Add message', new_capsule_message_path(@capsule), class: "btn btn-dark rounded-pill px-4 mb-3", data: { 'toggle-target': 'capsuleInput' } %>
+    <%= f.input :photos,
+                direct_upload: true,
+                required: true,
+                autofocus: true,
+                as: :file,
+                wrapper_html: { data: { 'toggle-target': 'capsuleInput' } },
+                input_html: { multiple: true } %>
+    <%= f.input :videos,
+                required: true,
+                autofocus: true,
+                as: :file,
+                wrapper_html: { data: { 'toggle-target': 'capsuleInput' } },
+                input_html: { multiple: true} %>
+    <%= f.input :audios,
+                required: true,
+                autofocus: true,
+                as: :file,
+                wrapper_html: { data: { 'toggle-target': 'capsuleInput' } },
+                input_html: { multiple: true } %>
+    <div data-controller="recipients" data-toggle-target="capsuleInput" class="d-none">
+        <label for="email_field" class="form-label mb-2">Recipient(s) email(s)</label>
+        <div class="d-flex w-100">
+          <input type="email" data-recipients-target="input", class="form-control w-50 me-4">
+          <button data-action="click->recipients#add", class="btn btn-dark rounded-pill px-4">Add Email</button>
+        </div>
+        <ul data-recipients-target="list" data-action="click->recipients#delete">
+          <!-- Les e-mails ajoutés seront affichés ici -->
+        </ul>
+    </div>
+    <%= f.input :unlock_date,
+                  as: :date,
+                  required: true,
+                  html5: true,
+                  input_html: { data: { controller: "datepicker" }, value: @capsule.unlock_date },
+                  wrapper_html: { class:'d-none', data: { 'toggle-target': 'capsuleInput' } } %>
+    <%= f.button :submit, class: 'btn btn-dark rounded-pill px-4 d-none', data: { 'toggle-target': 'capsuleInput' } %>
+  <% end %>
+  <div class="d-flex justify-content-center w-100">
+    <button data-action="click->toggle#toggleInputs" class="btn btn-dark rounded-pill px-4" data-toggle-target="capsuleInput">Next</button>
+  </div>
+</div>
