--- app/views/shared/_form_capsule_page_1.html.erb
@@ -1,48 +1,0 @@
-<div class="container"&>
-  <%= link_to 'Back', capsules_path, method: :patch %>
-  <!-- Toggle message section -->
-  <div class="container mt-3">
-    <% @capsule.messages.each do |message| %>
-      <div class="form-control">
-        <p><%= message.title %></p>
-        <p><%= message.content %></p>
-        <%= link_to 'Edit', edit_capsule_message_path(@capsule, message) %>
-        <%= button_to 'Delete', capsule_message_path(@capsule, message), method: :delete %>
-      </div>
-    <% end %>
-  </div>
-  <!-- Toggle photos section -->
-  <div>
-    <% @capsule.photos.each do |photo| %>
-      <%# <%= render partial: "capsules/thumbnail", locals: { photo: photo } %> %>
-    <% end %>
-  </div>
-  <!-- Capsule form -->
-  <%= simple_form_for @capsule, html: { multipart: true } do |f| %>
-    <%= f.input :name, label: 'Capsule name', input_html: { 'data-toogle-target': 'capsuleInput' } %>
-    <%= link_to 'Add message', new_capsule_message_path(@capsule), class: "btn btn-dark rounded-pill px-4 mb-3" %>
-    <%= f.input :photos,
-                direct_upload: true,
-                required: true,
-                autofocus: true,
-                as: :file,
-                wrapper_html: { 'data-toogle-target': 'capsuleInput' },
-                input_html: { multiple: true } %>
-    <%= f.input :videos,
-                required: true,
-                autofocus: true,
-                as: :file,
-                wrapper_html: { 'data-toogle-target': 'capsuleInput' },
-                input_html: { multiple: true} %>
-    <%= f.input :audios,
-                required: true,
-                autofocus: true,
-                as: :file,
-                wrapper_html: { 'data-toogle-target': 'capsuleInput' },
-                input_html: { multiple: true } %>
-    <%= f.submit %>
-  <% end %>
-  <div class="d-flex justify-content-center w-100">
-    <button data-action="click->toggle#toggleInputs" class="btn btn-dark rounded-pill px-4" >Next</button>
-  </div>
-</div>
