--- app/views/shared/_form_capsule_page_2.html.erb
@@ -1,27 +1,0 @@
-<div data-toggle-target="partial" class="d-none container">
-  <h3>PAGE 2: SET THE RECIPIENTS + COUNTDOWN</h3>
-  <div class="container mt-3">
-    <button data-action="click->toggle#toggleCaspuleDataPartial" class="btn btn-dark rounded-pill px-4">Back</button>
-    <%= simple_form_for(@capsule) do |f| %>
-      <div data-controller="add-recipients-email">
-        <label for="email_field">Enter one or several recipient-s email-s:</label>
-        <input type="email" data-add-recipients-email-target="input">
-        <button data-action="click->add-recipients-email#add" >Add Email</button>
-          <ul data-add-recipients-email-target="list" data-action="click->add-recipients-email#delete">
-            <!-- Les e-mails ajoutés seront affichés ici -->
-          </ul>
-      </div>
-      <%#= f.input :email, as: :email, required: true %>
-      <%= f.input :unlock_date, 
-                  as: :date, 
-                  required: true,
-                  html5: true,
-                  input_html: {
-                    data: { controller: "datepicker" }
-                  }
-                   %>
-      <%= f.button :submit %>
-    <% end %>
-
-  </div>
-</div>
